# swagger.yaml
openapi: "3.0.3"
info:
  title: "Kalyanipura Church API"
  description: "REST API for Kalyanipura Church React application"
  version: "1.0.0"
  contact:
    email: "hello@kalyanipurachurch.org"
servers:
  - url: "http://localhost:5000/api/v1"
    description: "Development server"
  - url: "https://api.kalyanipurachurch.org/api/v1"
    description: "Production server"
components:
  securitySchemes:
    BearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "Admin JWT token. Use POST /api/v1/admin/login to obtain."
  schemas:
    SuccessResponse:
      type: "object"
      properties:
        success:
          type: "boolean"
          example: true
        message:
          type: "string"
          example: "Success"
        data:
          type: "object"
    ErrorResponse:
      type: "object"
      properties:
        success:
          type: "boolean"
          example: false
        message:
          type: "string"
          example: "Error message"
        errors:
          type: "array"
          items:
            type: "object"
            properties:
              field:
                type: "string"
              message:
                type: "string"
...
paths:
...
  /admin/login:
    post:
      tags: ["Admin"]
      summary: "Admin login"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required: ["email", "password"]
              properties:
                email:
                  type: "string"
                  format: "email"
                password:
                  type: "string"
      responses:
        "200":
          description: "Login successful"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: "Invalid credentials"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...